+++ src/devices/unifi/index.ts
@@ -1,3 +1,4 @@
+export * from './own-unifi-air-quality-sensor';
 export * from './own-unifi-camera';
 export * from './unifi-protect';
 export * from './unifi-access';